Fixed issue with transferring same named files into workspace.
Was unable to transfer files with same filename but from different source directories into work directory.
Problem for tools like cuffquant which all have same named output (abundances.cxb).
Now check to see if input file would overwrite an existing file and if so add a randomer tag to the beginning of the filename.