Added Randomer tag to CombineGVCF output.
Fixes issue of gvcf and gvcf_idx needing same randomer added when being transferred to separate instances.
The problem is that merging multisample output causes file collision for downstream pipelines.
To fix this, we add a randomer tag to the beginning of each file name if it already exists on the instance.
Normally this is fine, but output from CombineGVCF produces an index file that must include the vcf as a suffix.